Janitor patch: Remove trailing whitespace from lines
Editors should remove them when saving, and most do, but that
tends to clutter up the diff output, so let's purge the evilness
in a single commit instead.

We only touch .c and .h files though, since whitespace has
semantic meaning for Makefiles and is useful to be able to add
to test-files and suchlike.

Signed-off-by: Andreas Ericsson <ae@op5.se>