Stop shadowing variables, types and functions
It's most often done with variables that are of the same type,
which means renaming them will cause some variable accesses to
access the global variable while others may use the local one.
Such bugs are generally nearly impossible to track down, since
one has to follow the actual logic to figure out which variable
is intended to be used (after all; The whole point of global vars
is that they can be used from anywhere).

Detected with -Wshadow option to gcc.

Signed-off-by: Andreas Ericsson <ae@op5.se>