Checks: Escape command line arguments correctly
Command line arguments should be stripped of a couple characters, so it isn't
possible to break out of for example strings, and start redirecting output and
write to files, given values of macros. And extra important is for macros that
can be updated from a user.

Previously, this wasn't that big issue, when macros itself could say that they
should be escaped. But the concept of escaping also depending which macro it
entered gives lot of other problems (notable for url:s at least, see previous
commit).

Only the requestor of the macro expansion knows where it's going to be used, so
only the requestor has the information to tell what to escape, and therefore,
this patch tells the macro expansion to escape the macros to be used as a
command line for check commands...

Signed-off-by: Max Sikström <msikstrom@op5.com>